Feature: opt-out - CloudFoundryApi: Implementation of bindServiceInstance - AutosleepServiceInstanceService: list of applications of the instanciated space, then call the new implemented method bindServiceInstance - Add tests on bindServiceInstance - Update space on bindServiceInstance - Update of application remote information by app checker - Corrections on debug ihm - Add delete application on ihm - Add task that bind applications every 24 hours - add exclude names parameter - add noOptOut parameter (not on IHM for now)

+++ src/main/java/org/cloudfoundry/autosleep/servicebroker/service/AutoSleepServiceInstanceBindingService.java
@@ -3,8 +3,10 @@ package org.cloudfoundry.autosleep.servicebroker.service;
 import lombok.extern.slf4j.Slf4j;
 import org.cloudfoundry.autosleep.dao.model.ApplicationBinding;
 import org.cloudfoundry.autosleep.dao.model.ApplicationInfo;
+import org.cloudfoundry.autosleep.dao.model.AutosleepServiceInstance;
 import org.cloudfoundry.autosleep.dao.repositories.ApplicationRepository;
 import org.cloudfoundry.autosleep.dao.repositories.BindingRepository;
+import org.cloudfoundry.autosleep.dao.repositories.ServiceRepository;
 import org.cloudfoundry.autosleep.scheduling.GlobalWatcher;
 import org.cloudfoundry.community.servicebroker.exception.ServiceBrokerException;
 import org.cloudfoundry.community.servicebroker.exception.ServiceInstanceBindingExistsException;
@@ -24,14 +26,18 @@ public class AutoSleepServiceInstanceBindingService implements ServiceInstanceBi
 
     private ApplicationRepository appRepository;
 
+    private ServiceRepository serviceRepository;
+
     private BindingRepository bindingRepository;
 
     private GlobalWatcher watcher;
 
     @Autowired
-    public AutoSleepServiceInstanceBindingService(ApplicationRepository appRepository, BindingRepository
-            bindingRepository, GlobalWatcher watcher) {
+    public AutoSleepServiceInstanceBindingService(ApplicationRepository appRepository,
+                                                  ServiceRepository serviceRepository,
+                                                  BindingRepository bindingRepository, GlobalWatcher watcher) {
         this.appRepository = appRepository;
+        this.serviceRepository = serviceRepository;
         this.bindingRepository = bindingRepository;
         this.watcher = watcher;
     }
@@ -64,17 +70,23 @@ public class AutoSleepServiceInstanceBindingService implements ServiceInstanceBi
     }
 
     @Override
-    public ServiceInstanceBinding deleteServiceInstanceBinding(DeleteServiceInstanceBindingRequest request) throws
-            ServiceBrokerException {
+    public ServiceInstanceBinding deleteServiceInstanceBinding(DeleteServiceInstanceBindingRequest request)
+            throws ServiceBrokerException {
         String bindingId = request.getBindingId();
-        log.debug("deleteServiceInstanceBinding - {}", bindingId);
-        ApplicationBinding binding = bindingRepository.findOne(bindingId);
-        log.debug("deleteServiceInstanceBinding on app ", binding.getAppGuid());
-        ApplicationInfo appInfo = appRepository.findOne(binding.getAppGuid());
-        appInfo.getStateMachine().onOptOut();
-        bindingRepository.delete(bindingId);
-        appRepository.save(appInfo);
-        //task launched will cancel by itself
-        return binding;
+        AutosleepServiceInstance serviceInstance = serviceRepository
+                .findOne(request.getInstance().getServiceInstanceId());
+        if (serviceInstance.isNoOptOut()) {
+            throw new ServiceBrokerException("unbinding this service is forbidden");
+        } else {
+            log.debug("deleteServiceInstanceBinding - {}", bindingId);
+            ApplicationBinding binding = bindingRepository.findOne(bindingId);
+            log.debug("deleteServiceInstanceBinding on app ", binding.getAppGuid());
+            ApplicationInfo appInfo = appRepository.findOne(binding.getAppGuid());
+            appInfo.getStateMachine().onOptOut();
+            bindingRepository.delete(bindingId);
+            appRepository.save(appInfo);
+            //task launched will cancel by itself
+            return binding;
+        }
     }
 }
